+++ dmr/settings.py
@@ -95,11 +95,15 @@ class HttpSpec(enum.StrEnum):
             like ``GET`` and ``HEAD`` can't have request bodies.
         empty_response_body: Disables validation that some status codes
             like ``204`` must not have response bodies.
+        header_name_server_managed: Disables validation that headers
+            like ``Server``, ``Keep-Alive``, etc. can't be
+            included in response headers.
 
     """
 
     empty_request_body = 'empty_request_body'
     empty_response_body = 'empty_response_body'
+    header_name_server_managed = 'header_name_server_managed'
 
 
 class SettingsDict(TypedDict, total=False):