+++ dmr/validation/endpoint_metadata.py
@@ -1,10 +1,17 @@
 import dataclasses
 import inspect
 import warnings
-from collections.abc import Callable, Sequence, Set
+from collections.abc import Callable, ItemsView, Sequence, Set
 from http import HTTPMethod, HTTPStatus
 from types import NoneType
-from typing import TYPE_CHECKING, Any, ClassVar, Final, TypeVar, assert_never
+from typing import (
+    TYPE_CHECKING,
+    Any,
+    ClassVar,
+    Final,
+    TypeVar,
+    assert_never,
+)
 
 from django.contrib.admindocs.utils import parse_docstring
 from django.core.cache.backends import dummy, locmem
@@ -44,6 +51,22 @@ from dmr.validation.payload import (
 if TYPE_CHECKING:
     from dmr.controller import Controller
     from dmr.errors import AsyncErrorHandler, SyncErrorHandler
+
+#: HTTP headers that are connection-specific or
+#: normally managed by the server.
+#: See RFC 9110 for more details.
+_FORBIDDEN_RESPONSE_HEADERS: Final = frozenset((
+    'connection',
+    'keep-alive',
+    'proxy-authenticate',
+    'proxy-authorization',
+    'te',
+    'trailer',
+    'transfer-encoding',
+    'upgrade',
+    'date',
+    'server',
+))
 
 #: HTTP methods that should not have a request body according to HTTP spec.
 #: These methods are: GET, HEAD, DELETE, CONNECT, TRACE.
@@ -141,6 +164,12 @@ class _ResponseListValidator:  # noqa: WPS214
             not in self.metadata.no_validate_http_spec
         ):
             self._check_empty_response_body(responses)
+        if (
+            HttpSpec.header_name_server_managed
+            not in self.metadata.no_validate_http_spec
+        ):
+            self._check_header_name_server_managed(responses)
+
         # TODO: add more checks
 
     def _check_empty_response_body(
@@ -164,11 +193,43 @@ class _ResponseListValidator:  # noqa: WPS214
                     f'with status code {response.status_code}',
                 )
 
+    def _check_header_name_server_managed(
+        self,
+        responses: list[ResponseSpec],
+    ) -> None:
+        endpoint_name = self.metadata.endpoint_name
+        for response in responses:
+            if not response.headers:
+                continue
+
+            forbidden_header = self._get_forbidden_header(
+                response.headers.items(),
+            )
+
+            if forbidden_header:
+                raise EndpointMetadataError(
+                    f'Header {forbidden_header!r} is not allowed in responses '
+                    f'from endpoint {endpoint_name!r}.',
+                )
+
     def _convert_responses(
         self,
         all_responses: list[ResponseSpec],
     ) -> dict[HTTPStatus, ResponseSpec]:
         return {resp.status_code: resp for resp in all_responses}
+
+    def _get_forbidden_header(
+        self,
+        response_headers: ItemsView[str, HeaderSpec],
+    ) -> str | None:
+
+        for header_name, header in response_headers:
+            if (
+                header_name.lower() in _FORBIDDEN_RESPONSE_HEADERS
+                and not header.skip_validation
+            ):
+                return header_name
+        return None
 
 
 @dataclasses.dataclass(slots=True, frozen=True, kw_only=True)